+++ apps/test-suite/TestModules.js
@@ -73,6 +73,10 @@ export function getTestModules() {
     optionalRequire(() => require('./tests/SQLiteLegacy'))
   );
 
+  if (['ios', 'web'].includes(Platform.OS)) {
+    modules.push(require('./tests/Fetch'));
+  }
+
   if (['android', 'ios'].includes(Platform.OS)) {
     modules.push(require('./tests/SQLite'));
   }